+++ 07.ls_object/lib/formatter/ls_formatter.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class LsFormatter
+  def initialize(entries)
+    @entries = entries
+  end
+end
